refactor: cashu wallet moved to cdk-common
