+++ crates/cdk-common/src/lib.rs
@@ -15,13 +15,13 @@ pub mod mint;
 pub mod payment;
 pub mod pub_sub;
 pub mod subscription;
+#[cfg(feature = "wallet")]
+pub mod wallet;
 pub mod ws;
 // re-exporting external crates
 pub use bitcoin;
 pub use cashu::amount::{self, Amount};
 pub use cashu::lightning_invoice::{self, Bolt11Invoice};
 pub use cashu::nuts::{self, *};
-#[cfg(feature = "wallet")]
-pub use cashu::wallet;
 pub use cashu::{dhke, ensure_cdk, mint_url, secret, util, SECP256K1};
 pub use error::Error;
